requirements.txt
streamlit
gTTS==2.2.2
googletrans==3.1.0a0
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
openai==0.28
tiktoken
paho-mqtt
Jinja2 >=2.9
contourpy >=1
numpy >=1.16
packaging >=16.8
pandas >=1.2
pillow >=7.1.0
PyYAML >=3.10
tornado >=5.1
xyzservices >=2021.09.1
bokeh==2.4.3
Streamlit-bokeh-events
